pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "make4db-snowflake"
authors = [
    {name = "Paresh Adhia", email = "padhia+github@gmail.com"},
]
description = "Snowflake provider for make4db tool"
readme = "README.md"
requires-python = ">=3.11"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "snowflake-snowpark-python",
    "sfconn",
    "make4db-api",
    "yappt",
]
dynamic = ["version"]

[tool.setuptools]
py-modules = ["make4db_snowflake"]

[tool.setuptools.dynamic]
version = {attr = "make4db_snowflake.__version__"}

[tool.ruff]
target-version = "py311"
line-length = 130
lint.extend-select = ["I"]

[tool.pytest.ini_options]
pythonpath = [ "." ]
filterwarnings = [ "ignore" ]
